dtos moved to dedicated project
