+++ WebShop/WebShop.Dto/Catalog/Category/DeleteCategoryDto.cs
@@ -3,7 +3,7 @@ using WebShop.Application.Common.Mappings;
 using WebShop.Application.MediatR.Catalog.Categories.Commands.DeleteCategory;
 using WebShop.Application.MediatR.Catalog.Categories.Commands.UpdateCategory;
 
-namespace WebShop.WebAPI.Models {
+namespace WebShop.Dto.Catalog.Category {
     public class DeleteCategoryDto : IMapWith<DeleteCategoryCommand> {
         public Guid? CategoryId { get; set; } = null!;
         public void Mapping(Profile profile) {
