+++ WebShop/WebShop.WebAPI/WebShop.WebAPI.csproj
@@ -21,6 +21,7 @@
   <ItemGroup>
     <ProjectReference Include="..\WebShop.Application\WebShop.Application.csproj" />
     <ProjectReference Include="..\WebShop.Domain\WebShop.Domain.csproj" />
+    <ProjectReference Include="..\WebShop.Dto\WebShop.Dto.csproj" />
     <ProjectReference Include="..\WebShop.Persistence\WebShop.Persistence.csproj" />
   </ItemGroup>
